Do not get child dt terms when filling geneVariant tw
