+++ client/tw/test/geneVariant.integration.spec.ts
@@ -2,6 +2,7 @@ import tape from 'tape'
 import type { GvCustomGsQ, RawGvTW, DtTerm } from '#types'
 import { vocabInit } from '#termdb/vocabulary'
 import { GvBase } from '../geneVariant'
+import { mayMakeGroups } from '../geneVariant.ts'
 
 /*************************
  reusable helper functions
@@ -18,7 +19,7 @@ async function getVocabApi() {
  test sections
 ***************/
 
-const vocabApi = await getVocabApi()
+const vocabApi: any = await getVocabApi()
 
 tape('\n', function (test) {
 	test.pass('-***- tw/geneVariant.integration -***-')
@@ -65,7 +66,6 @@ tape('fill() valid tw', async test => {
 	}
 	test.deepEqual(fullTw.q, expectedQ, 'should fill in q')
 	test.deepEqual(fullTw.term.groupsetting, { disabled: false }, 'should fill in term.groupsetting')
-	test.deepEqual(fullTw.term.childTerms, childTerms, 'should fill in term.childTerms')
 	test.end()
 })
 
@@ -87,6 +87,77 @@ tape('fill() tw with defaultQ', async test => {
 	test.deepEqual(fullTw.q, customGsQ, 'should fill in q with custom groupset')
 	test.deepEqual(fullTw.term.groupsetting, { disabled: false }, 'should fill in term.groupsetting')
 	test.deepEqual(fullTw.term.childTerms, childTerms, 'should fill in term.childTerms')
+	test.end()
+})
+
+tape('mayMakeGroups: fills groups', async test => {
+	const tw: any = {
+		q: { type: 'custom-groupset' },
+		term: {
+			kind: 'gene',
+			id: 'TP53',
+			gene: 'TP53',
+			name: 'TP53',
+			type: 'geneVariant'
+		}
+	}
+	await mayMakeGroups(tw, vocabApi)
+	test.equal(tw.q.customset.groups.length, 3, 'Should create 3 groups')
+	const grp1 = tw.q.customset.groups[1]
+	const grp2 = tw.q.customset.groups[2]
+	test.equal(grp1.name, 'Wildtype (somatic)', 'Group 1 name should be Wildtype (somatic)')
+	test.equal(grp2.name, 'SNV/indel (somatic)', 'Group 2 name should be SNV/indel (somatic)')
+	test.equal(grp1.filter.lst.length, 1, 'Group 1 filter.lst should have 1 item')
+	test.equal(grp2.filter.lst.length, 1, 'Group 2 filter.lst should have 1 item')
+	const grp1Item = grp1.filter.lst[0]
+	const grp2Item = grp2.filter.lst[0]
+	test.equal(grp1Item.type, 'tvs', 'Group 1 filter item should be a tvs')
+	test.equal(grp2Item.type, 'tvs', 'Group 2 filter item should be a tvs')
+	test.equal(grp1Item.tvs.values.length, 1, 'Group 1 tvs.values should have a single value')
+	test.equal(grp2Item.tvs.values.length, 1, 'Group 2 tvs.values should have a single value')
+	const grp1Value = grp1Item.tvs.values[0]
+	const grp2Value = grp2Item.tvs.values[0]
+	test.deepEqual(grp1Value, { key: 'WT', label: 'Wildtype', value: 'WT' }, 'Group 1 value should be a wildtype object')
+	test.deepEqual(grp2Value, { key: 'WT', label: 'Wildtype', value: 'WT' }, 'Group 2 value should be a wildtype object')
+	test.false(grp1Item.tvs.isnot, 'Group 1 tvs.isnot should be undefined')
+	test.true(grp2Item.tvs.isnot, 'Group 2 tvs.isnot should be true')
+	const grp0 = tw.q.customset.groups[0]
+	test.true(grp0.uncomputable, 'First group should be uncomputable')
+	test.end()
+})
+
+tape('mayMakeGroups: does nothing if groups already present', async test => {
+	const grp1 = { name: 'existing' }
+	const tw: any = {
+		q: { type: 'custom-groupset', customset: { groups: [grp1] } },
+		term: {
+			kind: 'gene',
+			id: 'TP53',
+			gene: 'TP53',
+			name: 'TP53',
+			type: 'geneVariant'
+		}
+	}
+	await mayMakeGroups(tw, vocabApi)
+	test.equal(tw.q.customset.groups.length, 1, 'Number of groups should be 1')
+	test.deepEqual(tw.q.customset.groups[0], grp1, 'Group 1 should equal grp1')
+	test.end()
+})
+
+tape('mayMakeGroups: does nothing if q.type != custom-groupset', async test => {
+	const tw: any = {
+		q: {},
+		term: {
+			kind: 'gene',
+			id: 'TP53',
+			gene: 'TP53',
+			name: 'TP53',
+			type: 'geneVariant'
+		}
+	}
+	const originalTw = structuredClone(tw)
+	await mayMakeGroups(tw, vocabApi)
+	test.deepEqual(tw, originalTw, 'tw should not change')
 	test.end()
 })
